Reverted changes rekated to Icon
